*Fix for ntfs writing *Cache size reduced which results in a bit slower writing but is more stable *Fix for the gameinstaller ProgressWindow *Added NTFS unmount on exit
WARNING:
The USBLoaderGX installer seems to be broken in some places. There are sometimes random freezes or crashes even on fat32. Those arent bad on a fat32 filesystem but can damage the whole partition on ntfs. Using the installer on ntfs is still very risky. You have been warned!

+++ source/libwbfs/libwbfs.h
@@ -67,7 +67,7 @@ typedef struct wbfs_disc_info
 
 // callback definition. Return 1 on fatal error (callback is supposed to make retries until no hopes..)
 typedef int (*rw_sector_callback_t)(void*fp,u32 lba,u32 count,void*iobuf);
-typedef void (*progress_callback_t)(u32 status, u32 total);
+typedef void (*progress_callback_t)(s64 status, s64 total);
 
 
 typedef struct wbfs_s
